Revert "Type of Scancode and Shiftcode -> uint (not uint16,uint32)"
This reverts commit 9097226bec72b8610d8135ebe80eef197a555eaf.

+++ deprecated.go
@@ -5,7 +5,7 @@ var OnWindowResize func(w, h uint)
 
 // Deprecated: use All()
 // Get all keyboard-event.
-func Full() (code rune, scan uint, shift uint) {
+func Full() (code rune, scan uint16, shift uint32) {
 	var flag uintptr = 0
 	if OnWindowResize != nil {
 		flag = ENABLE_WINDOW_INPUT
